+++ lib/src/database/commands/faction/imperials/pinned_down.dart
@@ -3,8 +3,8 @@ import 'package:swlegion/swlegion.dart';
 final pinnedDown = CommandCard(
   id: 'pinned-down',
   name: 'Pinned Down',
-  unitsActivated: '2 Support or Heavy Units',
-  factionRequired: Faction.imperials,
+  activated: '2 Support or Heavy Units',
+  faction: Faction.imperials,
   pips: 2,
   text: r'''
     After a friendly supoport or heavy unit performs an attack against a unit with a faceup order token, shuffle that token back into its order pool.