+++ lib/src/database/commands/faction/rebels/covering_fire.dart
@@ -3,8 +3,8 @@ import 'package:swlegion/swlegion.dart';
 final coveringFire = CommandCard(
   id: 'covering-fire',
   name: 'Covering Fire',
-  unitsActivated: '3 Corps Unit',
-  factionRequired: Faction.rebels,
+  activated: '3 Corps Unit',
+  faction: Faction.rebels,
   pips: 3,
   text: r'''
     After a friendly corps unit performs a ranged attack, another friendly unit at range 1-2 may gain 1 dodge token.