+++ lib/src/database/commands/neutral/ambush.dart
@@ -3,7 +3,7 @@ import 'package:swlegion/swlegion.dart';
 final ambush = CommandCard(
   id: 'ambush',
   name: 'Ambush',
-  unitsActivated: '1 Unit',
+  activated: '1 Unit',
   pips: 1,
   waves: const [
     Wave.wCore,