+++ lib/src/database/commands/neutral/assault.dart
@@ -3,7 +3,7 @@ import 'package:swlegion/swlegion.dart';
 final assault = CommandCard(
   id: 'assault',
   name: 'Assault',
-  unitsActivated: '3 Units',
+  activated: '3 Units',
   pips: 3,
   waves: const [
     Wave.wCore,