+++ lib/src/database/commands/units/boba_fett/whipcord_launcher.dart
@@ -4,8 +4,8 @@ import 'package:swlegion/swlegion.dart';
 final whipcordLauncher = CommandCard(
   id: 'whipcord-launcher',
   name: 'Whipcord Launcher',
-  unitsActivated: Units.bobaFett.name,
-  unitsRequired: [Units.bobaFett],
+  activated: Units.bobaFett.name,
+  required: [Units.bobaFett],
   pips: 1,
   text: r'''
     Boba Fett can perform moves while engaged with a unit that has 1 or more immobilized tokens. During Boba Fett's activation he gains: