+++ lib/src/database/commands/units/boba_fett/z_6_jetpack_launcher.dart
@@ -4,8 +4,8 @@ import 'package:swlegion/swlegion.dart';
 final z6JetpackLauncher = CommandCard(
   id: 'z-6-jetpack-launcher',
   name: 'Z-6 Jetpack Launcher',
-  unitsActivated: Units.bobaFett.name,
-  unitsRequired: [Units.bobaFett],
+  activated: Units.bobaFett.name,
+  required: [Units.bobaFett],
   pips: 3,
   text: r'''
     During Boba Fett's activation he gains the following weapon: