+++ lib/src/database/commands/units/bossk/lying_in_wait.dart
@@ -4,8 +4,8 @@ import 'package:swlegion/swlegion.dart';
 final lyingInWait = CommandCard(
   id: 'lying-in-wait',
   name: 'Lying In Wait',
-  unitsActivated: Units.bossk.name,
-  unitsRequired: [Units.bossk],
+  activated: Units.bossk.name,
+  required: [Units.bossk],
   pips: 3,
   text: r'''
     Bossk gains SHARPSHOOTER 1. After an enemy unit activates, Bossk gains 1