+++ lib/src/database/commands/units/bossk/reptilian_rampage.dart
@@ -4,8 +4,8 @@ import 'package:swlegion/swlegion.dart';
 final reptilianRampage = CommandCard(
   id: 'reptilian-rampage',
   name: 'Reptilian Rampage',
-  unitsActivated: Units.bossk.name,
-  unitsRequired: [Units.bossk],
+  activated: Units.bossk.name,
+  required: [Units.bossk],
   pips: 2,
   text: r'''
     During Bossk's Activation, he can perform up to 2 attack actions and he may