+++ lib/src/database/commands/units/emperor_palpatine.dart/an_entire_legion.dart
@@ -4,8 +4,8 @@ import 'package:swlegion/swlegion.dart';
 final anEntireLegion = CommandCard(
   id: 'an-entire-legion',
   name: 'An Entire Legion',
-  unitsActivated: 'Each Eligible Unit',
-  unitsRequired: [Units.emperorPalpatine],
+  activated: 'Each Eligible Unit',
+  required: [Units.emperorPalpatine],
   pips: 3,
   waves: const [
     Wave.wEmperorPalpatineCommanderExpansion,