+++ lib/src/database/commands/units/han_solo/reckless_diversion.dart
@@ -4,8 +4,8 @@ import 'package:swlegion/swlegion.dart';
 final recklessDiversion = CommandCard(
   id: 'reckless-diversion',
   name: 'Reckless Diversion',
-  unitsActivated: '${Units.hanSolo.name} and 1 Trooper',
-  unitsRequired: [Units.hanSolo],
+  activated: '${Units.hanSolo.name} and 1 Trooper',
+  required: [Units.hanSolo],
   pips: 2,
   text: r'''
     When an enemy unit performs an attack, it must attack a trooper unit that has a faceup order token, if available.