+++ lib/src/database/commands/units/jyn_erso/trust_goes_both_ways.dart
@@ -4,8 +4,8 @@ import 'package:swlegion/swlegion.dart';
 final trustGoesBothWays = CommandCard(
   id: 'trust-goes-both-ways',
   name: 'Trust Goes Both Ways',
-  unitsActivated: '${Units.jynErso.name} & 1 Trooper',
-  unitsRequired: [Units.jynErso],
+  activated: '${Units.jynErso.name} & 1 Trooper',
+  required: [Units.jynErso],
   pips: 2,
   text: r'''
     Jyn Erso gains INSPIRE 2.