+++ lib/src/database/commands/units/leia_organa/no_time_for_sorrows.dart
@@ -4,8 +4,8 @@ import 'package:swlegion/swlegion.dart';
 final noTimeForSorrows = CommandCard(
   id: 'no-time-for-sorrows',
   name: 'No Time for Sorrows',
-  unitsActivated: '2 Troopers',
-  unitsRequired: [Units.leiaOrgana],
+  activated: '2 Troopers',
+  required: [Units.leiaOrgana],
   pips: 2,
   text: r'''
     When a friendly Trooper unit is issued an order, it may perform a speed-1 move.